Create .NET Standard Class Library - Break project into a .net standard class library and a benchmark utility separately - This will allow other .NET apps to leverage this library and it could eventually be published to NuGet as an assembly
