+++ PruningRadixTrie/PruningRadixTrie.Benchmark.csproj
@@ -9,4 +9,8 @@
     <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\PruingRadixTrie\PruingRadixTrie.csproj" />
+  </ItemGroup>
+
 </Project>
